dev-requirements.txt
flake8    # for the travis build
pytest
pytest_ckan
pytest-cov
